toolkit/components/passwordmgr/test/mochitest/test_autocomplete_https_downgrade.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test autocomplete on an HTTPS page using upgraded HTTP logins</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="../../../satchel/test/satchel_common.js"></script>
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<script>
const chromeScript = runChecksAfterCommonInit(false);

runInParent(function addLogins() {
  const {Services} = ChromeUtils.import("resource://gre/modules/Services.jsm");

  // Create some logins just for this form, since we'll be deleting them.
  let nsLoginInfo = Components.Constructor("@mozilla.org/login-manager/loginInfo;1",
                                           Ci.nsILoginInfo, "init");

  // We have two actual HTTPS to avoid autofill before the schemeUpgrades pref flips to true.
  let login0 = new nsLoginInfo("https://example.com", "https://example.com", null,
                               "name", "pass", "uname", "pword");

  let login1 = new nsLoginInfo("https://example.com", "https://example.com", null,
                               "name1", "pass1", "uname", "pword");

  // Same as above but HTTP instead of HTTPS (to test de-duping)
  let login2 = new nsLoginInfo("http://example.com", "http://example.com", null,
                               "name1", "pass1", "uname", "pword");

  // Different HTTP login to upgrade with secure formActionOrigin
  let login3 = new nsLoginInfo("http://example.com", "https://example.com", null,
                               "name2", "passHTTPtoHTTPS", "uname", "pword");

  try {
    Services.logins.addLogin(login0);
    Services.logins.addLogin(login1);
    Services.logins.addLogin(login2);
    Services.logins.addLogin(login3);
  } catch (e) {
    assert.ok(false, "addLogin threw: " + e);
  }
});
</script>
<p id="display"></p>

<!-- we presumably can't hide the content for this test. -->
<div id="content">
  <iframe></iframe>
</div>

<pre id="test">
<script class="testbody" type="text/javascript">
let iframe = document.getElementsByTagName("iframe")[0];
let iframeDoc, hostname;
let uname;
let pword;

// Restore the form to the default state.
function restoreForm() {
  return SpecialPowers.spawn(getIframeBrowsingContext(window), [], function() {
    this.content.document.getElementById("form-basic-password").focus();
    this.content.document.getElementById("form-basic-username").value = "";
    this.content.document.getElementById("form-basic-password").value = "";
    this.content.document.getElementById("form-basic-username").focus();
  });
}

const HTTP_FORM_URL = "http://example.com/tests/toolkit/components/passwordmgr/test/mochitest/form_basic.html";

async function setup(formUrl) {
  await SpecialPowers.pushPrefEnv({"set": [["signon.schemeUpgrades", true]]});

  iframe.src = formUrl;
  let processedPromise = promiseFormsProcessed();
  await new Promise(resolve => {
    iframe.addEventListener("load", function() {
      resolve();
    }, {once: true});
  });

  await processedPromise;

  hostname = await SpecialPowers.spawn(getIframeBrowsingContext(window), [], function() {
    return this.content.document.documentURIObject.host;
  });
}

add_task(async function test_autocomplete_https_downgrade() {
  info("test_autocomplete_http, setup with " + HTTP_FORM_URL);
  await setup(HTTP_FORM_URL);

  LoginManager.getAllLogins().then(logins => {
    info("got logins: " + logins.map(l => l.origin));
  });
  // from a HTTP page, look for matching logins, we should never offer a login with an HTTPS scheme
  // we're expecting just login2 as a match
  // Make sure initial form is empty.
  await checkLoginFormInChildFrame(iframe, "form-basic-username", "name1", "form-basic-password", "pass1");

  // Trigger autocomplete popup
  await restoreForm();
  let popupState = await getPopupState();
  is(popupState.open, false, "Check popup is initially closed");
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown");
  let results = await shownPromise;
  info("got results: " + results.join(", "));
  popupState = await getPopupState();
  is(popupState.selectedIndex, -1, "Check no entries are selected");
  checkAutoCompleteResults(results, ["name1", "name2"], hostname, "initial");
});
</script>
</pre>
</body>
</html>
